pyproject.toml
[tool.poetry]
name = "fmp-py"
version = "0.0.18.0"
description = "Python package for Financial Modeling Prep API"
authors = ["TexasCoding <jeff10278@me.com>"]
readme = "README.md"

homepage = "https://pypi.org/project/fmp-py/"
repository = "https://github.com/TexasCoding/fmp-py"
documentation = "https://fmp-py.readthedocs.io/en/latest/index.html"

packages = [{include = "fmp_py", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
requests-mock = "^1.12.1"
pendulum = "^3.0.0"
ta = "^0.11.0"
requests-cache = "^1.2.1"
requests-ratelimiter = "^0.7.0"


[tool.poetry.group.tests.dependencies]
pytest = "^8.2.2"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.2"
pre-commit = "^3.7.1"
ipykernel = "^6.29.5"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.4.7"
sphinx-autobuild = "^2024.4.16"
myst-parser = "^3.0.1"
nbsphinx = "^0.9.4"
sphinx-autoapi = "^3.2.0"
sphinx-rtd-theme = "^2.0.0"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
